\docType{package}
\name{rk.CohenKappa-package}
\alias{rk.CohenKappa}
\alias{rk.CohenKappa-package}
\title{The rk.CohenKappa Package}
\description{
  RKWard GUI to calculate Cohen's Kappa.
}
\details{
  \tabular{ll}{ Package: \tab rk.CohenKappa\cr Type: \tab
  Package\cr Version: \tab 0.01-4\cr Date: \tab
  2013-02-24\cr Depends: \tab rkward (>= 0.5.6)\cr
  Enhances: \tab rkward\cr License: \tab GPL (>= 3)\cr
  LazyLoad: \tab yes\cr URL: \tab http://rkward.sf.net\cr }

  RKWard GUI to calculate Cohen's Kappa, i.e., the
  agreement of two raters (using the psych package)
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

