% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rk.CohenKappa-package.R
\docType{package}
\name{rk.CohenKappa-package}
\alias{rk.CohenKappa-package}
\title{The rk.CohenKappa Package}
\description{
RKWard GUI to calculate Cohen's Kappa.
}
\details{
\tabular{ll}{
Package: \tab rk.CohenKappa\cr
Type: \tab Package\cr
Version: \tab 0.01-6\cr
Date: \tab 2015-03-13\cr
Depends: \tab rkward (>= 0.6.0)\cr
Enhances: \tab rkward\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://rkward.sf.net\cr
}

RKWard GUI to calculate Cohen's Kappa, i.e.,
      the agreement of two raters (using the psych package)
}
\author{
Meik Michalke
}
\keyword{package}

